+++ package.json
@@ -105,6 +105,7 @@
     "react-datetime": "^2.14.0",
     "react-dom": "^16.4.0",
     "react-dropzone": "^4.2.11",
+    "react-google-maps": "^9.4.5",
     "react-helmet": "^5.2.0",
     "react-hot-loader": "^4.3.3",
     "react-intl": "^2.4.0",
