+++ src/pages/Discover/Components/StationMarker.scss
@@ -1,0 +1,31 @@
+.stationInfoWrapper {
+  width: 213px;
+  background: #FFFFFF;
+  box-shadow: 0 2px 8px 0 rgba(170,170,170,0.50);
+  border-radius: 5px;
+  padding: 8px;
+}
+
+.stationInfo {
+  font-family: motiva-sans, sans-serif;
+  letter-spacing: 0;
+  .s-name {
+    font-weight: 600;
+    font-size: 18px;
+    color: #484848;
+    line-height: 24px;
+  }
+  hr {
+    margin-top: 8px !important;
+    margin-bottom: 8px !important;
+  }
+  .s-label {
+    font-size: 15px;
+    color: #868686;
+  }
+  .s-value {
+    font-weight: 600;
+    font-size: 16px;
+    color: #484848;
+  }
+}
